VWV - Checkbox PartikelZufall und Feinjustierungen FlowField

+++ SlideShowTransition.VomWindeVerweht/Shader/MosaikShader.hlsl
@@ -13,7 +13,7 @@ struct PartikelDaten
     int lod;
     int status;
     
-    int gewicht;
+    float gewicht;
 };
 
 StructuredBuffer<PartikelDaten> PartikelBuffer : register(t0);
